Move Grid to separate file
